firestore_updater_cloud_function/requirements.txt
requests==2.28.2
functions-framework==3.*
google-cloud-firestore==2.9.1
google-cloud-storage==2.7.0